refactor(web): 提词补全过滤抽为共享模块，AutocompletePanel 支持向下展开
- 新增 utils/autocomplete.ts：filterAndScore / detectAcMode 共享过滤与触发映射
- ChatInput acFiltered 改用共享 filterAndScore（行为不变）
- AutocompletePanel 新增 placement=bottom，供表单场景面板向下展开
